+++ src/TRoom.h
@@ -53,10 +53,10 @@ public:
     int environment;
     int weight;
     bool isLocked;
-    int min_x;
-    int min_y;
-    int max_x;
-    int max_y;
+    qreal min_x;
+    qreal min_y;
+    qreal max_x;
+    qreal max_y;
 
     qint8 c;
     QString name;
